Rename super class to Campaign and moved it to be under /bot
